requirements.txt.txt
streamlit==1.40.2
torch==2.5.1
torchvision==0.20.1
Pillow==10.4.0